# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "tdPSA API Documentation",
        "description": "API documentation for tdPSA Clients module",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "tdPSA API Server"
        }
    ],
    "paths": {
...
        "/api/v1/risk/assessments": {
            "get": {
                "tags": [
                    "Risk"
                ],
                "summary": "Get list of risk assessments",
                "operationId": "getRiskAssessmentList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Missing risk.read scope"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Risk"
                ],
                "summary": "Create risk assessment",
                "operationId": "createRiskAssessment",
                "responses": {
                    "201": {
                        "description": "Assessment created"
                    },
                    "403": {
                        "description": "Missing risk.create scope"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/risk/assessments/{assessment}": {
            "get": {
                "tags": [
                    "Risk"
                ],
                "summary": "Get risk assessment",
                "operationId": "getRiskAssessmentById",
                "parameters": [
                    {
                        "name": "assessment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Missing risk.read scope"
                    },
                    "404": {
                        "description": "Assessment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Risk"
                ],
                "summary": "Update risk assessment",
                "operationId": "updateRiskAssessment",
                "parameters": [
                    {
                        "name": "assessment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assessment updated"
                    },
                    "403": {
                        "description": "Missing risk.update scope"
                    },
                    "404": {
                        "description": "Assessment not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/risk/assessments/{assessment}/items": {
            "post": {
                "tags": [
                    "Risk"
                ],
                "summary": "Create risk item",
                "operationId": "createRiskItem",
                "parameters": [
                    {
                        "name": "assessment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Risk item created"
                    },
                    "403": {
                        "description": "Missing risk.create scope"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/risk/items/{item}": {
            "get": {
                "tags": [
                    "Risk"
                ],
                "summary": "Get risk item",
                "operationId": "getRiskItemById",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Missing risk.read scope"
                    },
                    "404": {
                        "description": "Risk item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Risk"
                ],
                "summary": "Update risk item details",
                "operationId": "updateRiskItem",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Risk item updated"
                    },
                    "403": {
                        "description": "Missing risk.update scope"
                    },
                    "404": {
                        "description": "Risk item not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/risk/items/{item}/updates": {
            "post": {
                "tags": [
                    "Risk"
                ],
                "summary": "Create risk item update",
                "operationId": "createRiskItemUpdate",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Risk item update created"
                    },
                    "403": {
                        "description": "Missing risk.update scope"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}